Add more test coverage.
