+++ lib/dwh/ddl/ddl_bq_test.go
@@ -207,30 +207,6 @@ func (d *DDLTestSuite) TestAlterTableAddColumnsSomeAlreadyExist() {
 	}
 }
 
-func (d *DDLTestSuite) TestAlterTableCreateTable() {
-	fqName := "mock_dataset.mock_table"
-	d.bigQueryStore.GetConfigMap().AddTableToConfig(fqName, types.NewDwhTableConfig(typing.Columns{}, nil, true, true))
-
-	ctx := context.Background()
-	tc := d.bigQueryStore.GetConfigMap().TableConfig(fqName)
-
-	alterTableArgs := ddl.AlterTableArgs{
-		Dwh:         d.bigQueryStore,
-		Tc:          tc,
-		FqTableName: fqName,
-		CreateTable: tc.CreateTable,
-		ColumnOp:    constants.Add,
-	}
-
-	err := ddl.AlterTable(ctx, alterTableArgs, typing.Column{Name: "name", KindDetails: typing.String})
-	assert.Equal(d.T(), 1, d.fakeBigQueryStore.ExecCallCount())
-
-	query, _ := d.fakeBigQueryStore.ExecArgsForCall(0)
-	assert.Equal(d.T(), query, fmt.Sprintf("CREATE TABLE IF NOT EXISTS %s (name string)", fqName), query)
-	assert.NoError(d.T(), err, err)
-	assert.Equal(d.T(), false, d.bigQueryStore.GetConfigMap().TableConfig(fqName).CreateTable)
-}
-
 func (d *DDLTestSuite) TestAlterTableDropColumnsBigQuerySafety() {
 	ctx := context.Background()
 	ts := time.Now()
