+++ lib/dwh/ddl/ddl_sflk_test.go
@@ -15,95 +15,6 @@ import (
 	"github.com/artie-labs/transfer/lib/typing/ext"
 )
 
-func (d *DDLTestSuite) TestCreateTable() {
-	type _testCase struct {
-		name string
-		cols []typing.Column
-
-		expectedQuery string
-	}
-
-	var (
-		happyPathCols = []typing.Column{
-			{
-				Name:        "user_id",
-				KindDetails: typing.String,
-			},
-		}
-		twoCols = []typing.Column{
-			{
-				Name:        "user_id",
-				KindDetails: typing.String,
-			},
-			{
-				Name:        "enabled",
-				KindDetails: typing.Boolean,
-			},
-		}
-		bunchOfCols = []typing.Column{
-			{
-				Name:        "user_id",
-				KindDetails: typing.String,
-			},
-			{
-				Name:        "enabled_boolean",
-				KindDetails: typing.Boolean,
-			},
-			{
-				Name:        "array",
-				KindDetails: typing.Array,
-			},
-			{
-				Name:        "struct",
-				KindDetails: typing.Struct,
-			},
-		}
-	)
-
-	testCases := []_testCase{
-		{
-			name:          "happy path",
-			cols:          happyPathCols,
-			expectedQuery: "CREATE TABLE IF NOT EXISTS demo.public.experiments (user_id string)",
-		},
-		{
-			name:          "happy path + enabled",
-			cols:          twoCols,
-			expectedQuery: "CREATE TABLE IF NOT EXISTS demo.public.experiments (user_id string,enabled boolean)",
-		},
-		{
-			name:          "complex table creation",
-			cols:          bunchOfCols,
-			expectedQuery: "CREATE TABLE IF NOT EXISTS demo.public.experiments (user_id string,enabled_boolean boolean,array array,struct variant)",
-		},
-	}
-
-	for index, testCase := range testCases {
-		ctx := context.Background()
-		fqTable := "demo.public.experiments"
-		d.snowflakeStore.GetConfigMap().AddTableToConfig(fqTable, types.NewDwhTableConfig(typing.Columns{}, nil, true, true))
-		tc := d.snowflakeStore.GetConfigMap().TableConfig(fqTable)
-
-		alterTableArgs := ddl.AlterTableArgs{
-			Dwh:         d.snowflakeStore,
-			Tc:          tc,
-			FqTableName: fqTable,
-			CreateTable: tc.CreateTable,
-			ColumnOp:    constants.Add,
-			CdcTime:     time.Now().UTC(),
-		}
-
-		err := ddl.AlterTable(ctx, alterTableArgs, testCase.cols...)
-		assert.NoError(d.T(), err, testCase.name)
-
-		execQuery, _ := d.fakeSnowflakeStore.ExecArgsForCall(index)
-		assert.Equal(d.T(), testCase.expectedQuery, execQuery, testCase.name)
-
-		// Check if the table is now marked as created where `CreateTable = false`.
-		assert.Equal(d.T(), d.snowflakeStore.GetConfigMap().TableConfig(fqTable).CreateTable,
-			false, d.snowflakeStore.GetConfigMap().TableConfig(fqTable), testCase.name)
-	}
-}
 func (d *DDLTestSuite) TestAlterComplexObjects() {
 	ctx := context.Background()
 	// Test Structs and Arrays
